se arregla un tipo de dato
para mayor compatiilidad se cambia el tipo de datos

+++ POO3-TP1-CHAT/src/model/TABLES-DB-CHAT.sql
@@ -6,8 +6,8 @@
 --DROP TABLE IF EXISTS usuarios;
 
 CREATE TABLE IF NOT EXISTS usuarios (
-  id_usuario_PK bigserial NOT NULL,
-  nombre_usuario  VARCHAR NULL,
+  id_usuario_PK int NOT NULL,
+  nombre_usuario VARCHAR NULL,
   PRIMARY KEY (id_usuario_PK) 
 );
